+++ src/Twig/RequestIdExtension.php
@@ -1,4 +1,5 @@
-<?php
+<?php declare(strict_types=1);
+
 /*
  * This file is part of chrisguitarguy/request-id-bundle
 
@@ -34,7 +35,7 @@ final class RequestIdExtension extends \Twig_Extension
     /**
      * {@inheritdoc}
      */
-    public function getFunctions()
+    public function getFunctions() : array
     {
         return [
             new \Twig_SimpleFunction('request_id', [$this->idStorage, 'getRequestId']),
@@ -44,7 +45,7 @@ final class RequestIdExtension extends \Twig_Extension
     /**
      * {@inheritdoc}
      */
-    public function getName()
+    public function getName() : string
     {
         return 'chirsguitarguy_request_id';
     }